# docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
...
  httpx.JSendError:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: Failed to connect to database
        type: string
      status:
        example: error
        type: string
    type: object
  httpx.JSendFail:
    properties:
      data:
        additionalProperties: {}
        type: object
      status:
        example: fail
        type: string
    type: object
...
  models.CreateMerchantRequest:
    properties:
      address:
        example: Av. Insurgentes Sur 1234, Col. Del Valle
        type: string
      business_name:
        example: Tacos El Güero
        maxLength: 255
        minLength: 2
        type: string
      business_type:
        example: restaurant
        type: string
      city:
        example: Ciudad de México
        type: string
      email:
        example: contacto@tacoselguero.com
        type: string
      latitude:
        example: 19.432608
        maximum: 90
        minimum: -90
        type: number
      longitude:
        example: -99.133209
        maximum: 180
        minimum: -180
        type: number
      phone:
        example: "+525512345678"
        type: string
      postal_code:
        example: "03100"
        type: string
      state:
        example: CDMX
        type: string
    required:
    - address
    - business_name
    - business_type
    - city
    - latitude
    - longitude
    - phone
    - state
    type: object
  models.CreateNotificationRequest:
    properties:
      body:
        example: 'Tu pedido #1234 está en camino'
        type: string
      data:
        type: string
      notification_type:
        allOf:
        - $ref: '#/definitions/models.NotificationType'
        enum:
        - order_created
        - order_updated
        - order_assigned
        - order_in_transit
        - order_delivered
        - order_canceled
        - payment_received
        - payment_failed
        - driver_assigned
        - driver_nearby
        - general
        - promotional
        example: order_in_transit
      title:
        example: Pedido en camino
        maxLength: 255
        type: string
      user_id:
        example: d53b655c-e833-400e-b0e8-ee68ea18e2cc
        type: string
    required:
    - body
    - notification_type
    - title
    - user_id
    type: object
...
  models.Merchant:
    properties:
      address:
        example: Av. Insurgentes Sur 1234, Col. Del Valle
        type: string
      business_name:
        example: Tacos El Güero
        type: string
      business_type:
        example: restaurant
        type: string
      city:
        example: Ciudad de México
        type: string
      country:
        example: MX
        type: string
      created_at:
        example: "2025-01-15T10:00:00Z"
        type: string
      email:
        example: contacto@tacoselguero.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      latitude:
        example: 19.432608
        type: number
      longitude:
        example: -99.133209
        type: number
      phone:
        example: "+525512345678"
        type: string
      postal_code:
        example: "03100"
        type: string
      rating:
        example: 4.5
        type: number
      state:
        example: CDMX
        type: string
      status:
        enum:
        - active
        - inactive
        - suspended
        example: active
        type: string
      total_orders:
        example: 125
        type: integer
      updated_at:
        example: "2025-01-15T10:00:00Z"
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
  models.MerchantListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Merchant'
        type: array
      status:
        example: success
        type: string
    type: object
  models.MerchantResponse:
    properties:
      data:
        $ref: '#/definitions/models.Merchant'
      status:
        example: success
        type: string
    type: object
  models.Notification:
    properties:
      body:
        example: 'Tu pedido #1234 está en camino'
        type: string
      created_at:
        example: "2025-01-15T10:00:00Z"
        type: string
      data:
        type: string
      id:
        example: d53b655c-e833-400e-b0e8-ee68ea18e2cc
        type: string
      is_read:
        example: false
        type: boolean
      notification_type:
        allOf:
        - $ref: '#/definitions/models.NotificationType'
        enum:
        - order_created
        - order_updated
        - order_assigned
        - order_in_transit
        - order_delivered
        - order_canceled
        - payment_received
        - payment_failed
        - driver_assigned
        - driver_nearby
        - general
        - promotional
        example: order_in_transit
      read_at:
        example: "2025-01-15T10:30:00Z"
        type: string
      title:
        example: Pedido en camino
        type: string
      user_id:
        example: d53b655c-e833-400e-b0e8-ee68ea18e2cc
        type: string
    type: object
...
  models.NotificationResponse:
    properties:
      data:
        $ref: '#/definitions/models.Notification'
      status:
        example: success
        type: string
    type: object
  models.NotificationType:
    enum:
    - order_created
    - order_updated
    - order_assigned
    - order_in_transit
    - order_delivered
    - order_canceled
    - payment_received
    - payment_failed
    - driver_assigned
    - driver_nearby
    - general
    - promotional
    type: string
    x-enum-varnames:
    - NotificationTypeOrderCreated
    - NotificationTypeOrderUpdated
    - NotificationTypeOrderAssigned
    - NotificationTypeOrderInTransit
    - NotificationTypeOrderDelivered
    - NotificationTypeOrderCanceled
    - NotificationTypePaymentReceived
    - NotificationTypePaymentFailed
    - NotificationTypeDriverAssigned
    - NotificationTypeDriverNearby
    - NotificationTypeGeneral
    - NotificationTypePromotional
...
  models.UpdateMerchantRequest:
    properties:
      address:
        example: Av. Insurgentes Sur 1234, Col. Del Valle
        type: string
      business_name:
        example: Tacos El Güero
        maxLength: 255
        minLength: 2
        type: string
      email:
        example: contacto@tacoselguero.com
        type: string
      latitude:
        example: 19.432608
        maximum: 90
        minimum: -90
        type: number
      longitude:
        example: -99.133209
        maximum: 180
        minimum: -180
        type: number
      phone:
        example: "+525512345678"
        type: string
      postal_code:
        example: "03100"
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Delivery marketplace API for customers, merchants, and drivers.
  title: TacoShare Delivery API
  version: "1.0"
paths:
...
  /merchants:
    get:
      consumes:
      - application/json
      description: List all merchants with optional filters
      parameters:
      - description: Filter by city
        in: query
        name: city
        type: string
      - description: Filter by business type
        in: query
        name: business_type
        type: string
      - description: Filter by status
        enum:
        - active
        - inactive
        - suspended
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Merchants retrieved successfully
          schema:
            $ref: '#/definitions/models.MerchantListResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/httpx.JSendError'
      summary: List merchants
      tags:
      - merchants
    post:
      consumes:
      - application/json
      description: Create a new merchant profile for the authenticated user
      parameters:
      - description: Merchant details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateMerchantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Merchant created successfully
          schema:
            $ref: '#/definitions/models.MerchantResponse'
        "400":
          description: Validation failed
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/httpx.JSendError'
      security:
      - BearerAuth: []
      summary: Create merchant
      tags:
      - merchants
  /merchants/{id}:
    get:
      consumes:
      - application/json
      description: Get merchant details by ID (admin only)
      parameters:
      - description: Merchant ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Merchant retrieved successfully
          schema:
            $ref: '#/definitions/models.MerchantResponse'
        "400":
          description: Invalid merchant ID
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "404":
          description: Merchant not found
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/httpx.JSendError'
      security:
      - BearerAuth: []
      summary: Get merchant by ID
      tags:
      - merchants
  /merchants/me:
    get:
      consumes:
      - application/json
      description: Get merchant profile for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Merchant retrieved successfully
          schema:
            $ref: '#/definitions/models.MerchantResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "404":
          description: Merchant not found
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/httpx.JSendError'
      security:
      - BearerAuth: []
      summary: Get my merchant
      tags:
      - merchants
    patch:
      consumes:
      - application/json
      description: Update merchant profile for the authenticated user
      parameters:
      - description: Merchant update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateMerchantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Merchant updated successfully
          schema:
            $ref: '#/definitions/models.MerchantResponse'
        "400":
          description: Validation failed
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "404":
          description: Merchant not found
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/httpx.JSendError'
      security:
      - BearerAuth: []
      summary: Update my merchant
      tags:
      - merchants
...
  /notifications/send:
    post:
      consumes:
      - application/json
      description: Send a push notification to a specific user. This endpoint creates
        the notification in the database and sends it via FCM to all active devices.
        Useful for testing and manual notifications.
      parameters:
      - description: Notification details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateNotificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Notification sent successfully
          schema:
            $ref: '#/definitions/models.NotificationResponse'
        "400":
          description: Validation failed
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "403":
          description: Forbidden - admin only
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/httpx.JSendError'
      security:
      - BearerAuth: []
      summary: Send notification (Admin)
      tags:
      - notifications-admin
...
produces:
- application/json
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"